- Work Programme Form
- WKP Application Process Implementation
- Merge Conflicts Fix
- Implementation of ;
- Implementation of GET & POST endpoints for "Minimum Requirement"
- Merge Conflicts Fix

+++ Backend_UMR_Work_Program/DataModels/WP_TOTAL_NO_OF_OIL_SPILLED_AND_RECOVERED_BY_YEAR_COMPANYNAME.cs
@@ -7,7 +7,7 @@ namespace Backend_UMR_Work_Program.Models
     {
         public string? CompanyName { get; set; }
         public string? Year_of_WP { get; set; }
-        public decimal? Total_Quantity_Spilled { get; set; }
-        public decimal? Total_Quantity_Recovered { get; set; }
+        public int? Total_Quantity_Spilled { get; set; }
+        public int? Total_Quantity_Recovered { get; set; }
     }
 }
